# [forge] Add warmup and cooldown to measuring performance