--- a/crates/transaction-emitter-lib/src/emitter/mod.rs
+++ b/crates/transaction-emitter-lib/src/emitter/mod.rs
@@ -5,10 +5,11 @@ pub mod account_minter;
 pub mod stats;
 pub mod submission_worker;
 
-use ::aptos_logger::*;
 use again::RetryPolicy;
 use anyhow::{anyhow, format_err, Result};
 use aptos_infallible::RwLock;
+use aptos_logger::sample::Sampling;
+use aptos_logger::{info, sample, sample::SampleRate, warn};
 use aptos_rest_client::Client as RestClient;
 use aptos_sdk::{
     move_types::account_address::AccountAddress,
@@ -33,7 +34,11 @@ use tokio::{runtime::Handle, task::JoinHandle, time};
 
 use crate::{
     args::TransactionType,
-    emitter::{account_minter::AccountMinter, submission_worker::SubmissionWorker},
+    emitter::{
+        account_minter::AccountMinter,
+        stats::{DynamicStatsTracking, TxnStats},
+        submission_worker::SubmissionWorker,
+    },
     transaction_generator::{
         account_generator::AccountGeneratorCreator, nft_mint::NFTMintGeneratorCreator,
         p2p_transaction_generator::P2PTransactionGeneratorCreator,
@@ -42,7 +47,6 @@ use crate::{
 };
 use aptos_sdk::transaction_builder::aptos_stdlib;
 use rand::rngs::StdRng;
-use stats::{StatsAccumulator, TxnStats};
 
 // Max is 100k TPS for a full day.
 const MAX_TXNS: u64 = 100_000_000_000;
@@ -74,6 +78,7 @@ pub struct EmitModeParams {
     pub start_jitter_millis: u64,
     pub wait_millis: u64,
     pub check_account_sequence_only_once_fraction: f32,
+    pub check_account_sequence_sleep_millis: u64,
 }
 
 #[derive(Clone, Debug)]
@@ -200,7 +205,7 @@ impl EmitJobRequest {
                 // The target mempool backlog is set to be 3x of the target TPS because of the on an average,
                 // we can ~3 blocks in consensus queue. As long as we have 3x the target TPS as backlog,
                 // it should be enough to produce the target TPS.
-                let transactions_per_account = 5;
+                let transactions_per_account = 20;
                 let num_workers_per_endpoint = max(
                     mempool_backlog / (clients_count * transactions_per_account),
                     1,
@@ -226,6 +231,7 @@ impl EmitJobRequest {
                     accounts_per_worker: 1,
                     workers_per_endpoint: num_workers_per_endpoint,
                     check_account_sequence_only_once_fraction: 0.0,
+                    check_account_sequence_sleep_millis: 300,
                 }
             }
             EmitJobMode::ConstTps { tps } => {
@@ -276,9 +282,16 @@ impl EmitJobRequest {
                     transactions_per_account, wait_seconds
                 );
 
+                // sample latency on 2% of requests, or at least once every 5s.
+                let sample_latency_fraction = 1.0_f32.min(0.02_f32.max(
+                    wait_seconds as f32
+                        / (clients_count * num_workers_per_endpoint) as f32
+                        / 5.0_f32,
+                ));
+
                 info!(
-                    " Will use {} clients and {} workers per client",
-                    clients_count, num_workers_per_endpoint
+                    " Will use {} clients and {} workers per client, sampling latency on {}",
+                    clients_count, num_workers_per_endpoint, sample_latency_fraction
                 );
 
                 EmitModeParams {
@@ -293,8 +306,8 @@ impl EmitJobRequest {
                     start_jitter_millis: 0,
                     accounts_per_worker: 1,
                     workers_per_endpoint: num_workers_per_endpoint,
-                    // sample latency on 2% of requests.
-                    check_account_sequence_only_once_fraction: 1.0 - 0.02,
+                    check_account_sequence_only_once_fraction: 1.0 - sample_latency_fraction,
+                    check_account_sequence_sleep_millis: 300,
                 }
             }
         }
@@ -310,7 +323,17 @@ struct Worker {
 pub struct EmitJob {
     workers: Vec<Worker>,
     stop: Arc<AtomicBool>,
-    stats: Arc<StatsAccumulator>,
+    stats: Arc<DynamicStatsTracking>,
+}
+
+impl EmitJob {
+    pub fn start_next_phase(&self) {
+        self.stats.start_next_phase();
+    }
+
+    pub fn get_cur_phase(&self) -> usize {
+        self.stats.get_cur_phase()
+    }
 }
 
 #[derive(Debug)]
@@ -349,6 +372,7 @@ impl TxnEmitter {
         &mut self,
         root_account: &mut LocalAccount,
         req: EmitJobRequest,
+        stats_tracking_phases: usize,
     ) -> Result<EmitJob> {
         let mode_params = req.calculate_mode_params();
         let workers_per_endpoint = mode_params.workers_per_endpoint;
@@ -369,7 +393,7 @@ impl TxnEmitter {
         let all_addresses = Arc::new(RwLock::new(all_addresses));
         let mut all_accounts = all_accounts.into_iter();
         let stop = Arc::new(AtomicBool::new(false));
-        let stats = Arc::new(StatsAccumulator::default());
+        let stats = Arc::new(DynamicStatsTracking::new(stats_tracking_phases));
         let tokio_handle = Handle::current();
         let txn_factory = self
             .txn_factory
@@ -463,7 +487,7 @@ impl TxnEmitter {
         })
     }
 
-    pub async fn stop_job(&mut self, job: EmitJob) -> TxnStats {
+    pub async fn stop_job(&mut self, job: EmitJob) -> Vec<TxnStats> {
         job.stop.store(true, Ordering::Relaxed);
         for worker in job.workers {
             let mut accounts = worker
@@ -472,23 +496,30 @@ impl TxnEmitter {
                 .expect("TxnEmitter worker thread failed");
             self.accounts.append(&mut accounts);
         }
+
         job.stats.accumulate()
     }
 
-    pub fn peek_job_stats(&self, job: &EmitJob) -> TxnStats {
+    pub fn peek_job_stats(&self, job: &EmitJob) -> Vec<TxnStats> {
         job.stats.accumulate()
     }
 
     pub async fn periodic_stat(&mut self, job: &EmitJob, duration: Duration, interval_secs: u64) {
         let deadline = Instant::now() + duration;
-        let mut prev_stats: Option<TxnStats> = None;
+        let mut prev_stats: Option<Vec<TxnStats>> = None;
+        let default_stats = TxnStats::default();
         let window = Duration::from_secs(max(interval_secs, 1));
         while Instant::now() < deadline {
             tokio::time::sleep(window).await;
+            let cur_phase = job.stats.get_cur_phase();
             let stats = self.peek_job_stats(job);
-            let delta = &stats - &prev_stats.unwrap_or_default();
+            let delta = &stats[cur_phase]
+                - prev_stats
+                    .as_ref()
+                    .map(|p| &p[cur_phase])
+                    .unwrap_or(&default_stats);
             prev_stats = Some(stats);
-            info!("{}", delta.rate(window));
+            info!("phase {}: {}", cur_phase, delta.rate(window));
         }
     }
 
@@ -498,13 +529,13 @@ impl TxnEmitter {
         emit_job_request: EmitJobRequest,
         duration: Duration,
     ) -> Result<TxnStats> {
-        let job = self.start_job(root_account, emit_job_request).await?;
+        let job = self.start_job(root_account, emit_job_request, 1).await?;
         info!("Starting emitting txns for {} secs", duration.as_secs());
         time::sleep(duration).await;
         info!("Ran for {} secs, stopping job...", duration.as_secs());
         let stats = self.stop_job(job).await;
         info!("Stopped job");
-        Ok(stats)
+        Ok(stats.into_iter().next().unwrap())
     }
 
     pub async fn emit_txn_for_with_stats(
@@ -515,12 +546,12 @@ impl TxnEmitter {
         interval_secs: u64,
     ) -> Result<TxnStats> {
         info!("Starting emitting txns for {} secs", duration.as_secs());
-        let job = self.start_job(root_account, emit_job_request).await?;
+        let job = self.start_job(root_account, emit_job_request, 1).await?;
         self.periodic_stat(&job, duration, interval_secs).await;
         info!("Ran for {} secs, stopping job...", duration.as_secs());
         let stats = self.stop_job(job).await;
         info!("Stopped job");
-        Ok(stats)
+        Ok(stats.into_iter().next().unwrap())
     }
 
     pub async fn submit_single_transaction(
@@ -553,9 +584,12 @@ async fn wait_for_single_account_sequence(
                 }
             }
             Err(e) => {
-                info!(
-                    "Failed to query sequence number for account {:?} for instance {:?} : {:?}",
-                    account, client, e
+                sample!(
+                    SampleRate::Duration(Duration::from_secs(60)),
+                    warn!(
+                        "Failed to query sequence number for account {:?} for instance {:?} : {:?}",
+                        account, client, e
+                    )
                 );
             }
         }
@@ -568,7 +602,7 @@ async fn wait_for_single_account_sequence(
 }
 
 /// This function waits for the submitted transactions to be committed, up to
-/// a deadline.
+/// a wait_timeout (counted from the start_time passed in, not from the function call).
 /// It returns number of transactions that expired without being committed,
 /// and sum of completion timestamps for those that have.
 ///
@@ -581,8 +615,8 @@ async fn wait_for_accounts_sequence(
     transactions_per_account: usize,
     wait_timeout: Duration,
     fetch_only_once: bool,
+    sleep_between_cycles: Duration,
 ) -> (usize, u128) {
-    let deadline = start_time + wait_timeout;
     let mut pending_addresses: HashSet<_> = accounts.iter().map(|d| d.address()).collect();
     let mut latest_fetched_counts = HashMap::new();
 
@@ -610,18 +644,21 @@ async fn wait_for_accounts_sequence(
                 }
             }
             Err(e) => {
-                info!(
-                    "Failed to query ledger info on accounts {:?} for instance {:?} : {:?}",
-                    pending_addresses, client, e
+                sample!(
+                    SampleRate::Duration(Duration::from_secs(60)),
+                    warn!(
+                        "Failed to query ledger info on accounts {:?} for instance {:?} : {:?}",
+                        pending_addresses, client, e
+                    )
                 );
             }
         }
 
-        if Instant::now() >= deadline {
+        if start_time.elapsed() >= wait_timeout {
             break;
         }
 
-        time::sleep(Duration::from_millis(1000)).await;
+        time::sleep(sleep_between_cycles).await;
     }
 
     (