--- a/crates/transaction-emitter-lib/src/emitter/stats.rs
+++ b/crates/transaction-emitter-lib/src/emitter/stats.rs
@@ -5,7 +5,7 @@ use std::{
     fmt,
     ops::Sub,
     sync::{
-        atomic::{AtomicU64, Ordering},
+        atomic::{AtomicU64, AtomicUsize, Ordering},
         Arc,
     },
     time::Duration,
@@ -29,6 +29,9 @@ pub struct TxnStatsRate {
     pub expired: u64,
     pub failed_submission: u64,
     pub latency: u64,
+    pub latency_samples: u64,
+    pub p50_latency: u64,
+    pub p90_latency: u64,
     pub p99_latency: u64,
 }
 
@@ -36,8 +39,8 @@ impl fmt::Display for TxnStatsRate {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(
             f,
-            "submitted: {} txn/s, committed: {} txn/s, expired: {} txn/s, failed submission: {} tnx/s, latency: {} ms, p99 latency: {} ms",
-            self.submitted, self.committed, self.expired, self.failed_submission, self.latency, self.p99_latency,
+            "submitted: {} txn/s, committed: {} txn/s, expired: {} txn/s, failed submission: {} tnx/s, latency: {} ms, (p50: {} ms, p90: {} ms, p99: {} ms), latency samples: {}",
+            self.submitted, self.committed, self.expired, self.failed_submission, self.latency, self.p50_latency, self.p90_latency, self.p99_latency, self.latency_samples,
         )
     }
 }
@@ -58,6 +61,9 @@ impl TxnStats {
             } else {
                 self.latency / self.latency_samples
             },
+            latency_samples: self.latency_samples,
+            p50_latency: self.latency_buckets.percentile(50, 100),
+            p90_latency: self.latency_buckets.percentile(90, 100),
             p99_latency: self.latency_buckets.percentile(99, 100),
         }
     }
@@ -114,8 +120,10 @@ impl StatsAccumulator {
     }
 }
 
-const DEFAULT_HISTOGRAM_CAPACITY: usize = 1024;
-const DEFAULT_HISTOGRAM_STEP_WIDTH: u64 = 50;
+// have more slots than generally used txn expiration. (240s)
+const DEFAULT_HISTOGRAM_CAPACITY: usize = 2400;
+// we don't have better precision than ~300 ms anyways.
+const DEFAULT_HISTOGRAM_STEP_WIDTH: u64 = 100;
 
 #[derive(Debug)]
 pub struct AtomicHistogramAccumulator {
@@ -155,17 +163,17 @@ impl AtomicHistogramAccumulator {
         }
     }
 
-    fn get_bucket_num(&self, data_value: u64) -> u64 {
+    fn get_bucket_num(&self, data_value: u64) -> usize {
         let bucket_num = data_value / self.step_width;
         if bucket_num >= self.capacity as u64 - 2 {
-            return self.capacity as u64 - 1;
+            return self.capacity - 1;
         }
-        bucket_num
+        bucket_num as usize
     }
 
     pub fn record_data_point(&self, data_value: u64, data_num: u64) {
         let bucket_num = self.get_bucket_num(data_value);
-        self.buckets[bucket_num as usize].fetch_add(data_num as u64, Ordering::Relaxed);
+        self.buckets[bucket_num].fetch_add(data_num as u64, Ordering::Relaxed);
     }
 }
 
@@ -220,6 +228,42 @@ impl AtomicHistogramSnapshot {
     }
 }
 
+#[derive(Debug)]
+pub struct DynamicStatsTracking {
+    num_phases: usize,
+    cur_phase: AtomicUsize,
+    stats: Vec<StatsAccumulator>,
+}
+
+impl DynamicStatsTracking {
+    pub fn new(num_phases: usize) -> DynamicStatsTracking {
+        assert!(num_phases >= 1);
+        Self {
+            num_phases,
+            cur_phase: AtomicUsize::new(0),
+            stats: (0..num_phases)
+                .map(|_| StatsAccumulator::default())
+                .collect(),
+        }
+    }
+
+    pub fn start_next_phase(&self) {
+        assert!(self.cur_phase.fetch_add(1, Ordering::Relaxed) + 1 < self.num_phases);
+    }
+
+    pub fn get_cur_phase(&self) -> usize {
+        self.cur_phase.load(Ordering::Relaxed)
+    }
+
+    pub fn get_cur(&self) -> &StatsAccumulator {
+        self.stats.get(self.get_cur_phase()).unwrap()
+    }
+
+    pub fn accumulate(&self) -> Vec<TxnStats> {
+        self.stats.iter().map(|s| s.accumulate()).collect()
+    }
+}
+
 #[cfg(test)]
 mod test {
     use crate::emitter::stats::{
@@ -238,10 +282,19 @@ mod test {
     pub fn test_get_bucket_num() {
         let histogram = AtomicHistogramAccumulator::default();
         assert_eq!(histogram.get_bucket_num(0), 0);
-        assert_eq!(histogram.get_bucket_num(49), 0);
-        assert_eq!(histogram.get_bucket_num(50), 1);
-        assert_eq!(histogram.get_bucket_num(51), 1);
-        assert_eq!(histogram.get_bucket_num(200_000), 1023);
+        assert_eq!(
+            histogram.get_bucket_num(DEFAULT_HISTOGRAM_STEP_WIDTH - 1),
+            0
+        );
+        assert_eq!(histogram.get_bucket_num(DEFAULT_HISTOGRAM_STEP_WIDTH), 1);
+        assert_eq!(
+            histogram.get_bucket_num(DEFAULT_HISTOGRAM_STEP_WIDTH + 1),
+            1
+        );
+        assert_eq!(
+            histogram.get_bucket_num(500_000),
+            DEFAULT_HISTOGRAM_CAPACITY - 1
+        );
     }
 
     #[test]