### Description

As metrics can differ during that period. I think keeping 5% on each end shouldn't reduce the test much, but might give useful stability to it.

### Test Plan
CI landblocking forge

Warmup stats: submitted: 7160 txn/s, committed: 7160 txn/s, expired: 0 txn/s, failed submission: 0 tnx/s, latency: 4082 ms, p99 latency: 6000 ms"}
Test stats: submitted: 6827 txn/s, committed: 6827 txn/s, expired: 0 txn/s, failed submission: 0 tnx/s, latency: 4380 ms, p99 latency: 6000 ms"}
Cooldown stats: submitted: 7042 txn/s, committed: 7042 txn/s, expired: 0 txn/s, failed submission: 0 tnx/s, latency: 4324 ms, p99 latency: 6000 ms"}

Bigger difference can be seen with ConstTps (3000):

Warmup stats: submitted: 2970 txn/s, committed: 2970 txn/s, expired: 0 txn/s, failed submission: 0 tnx/s, latency: 1381 ms, p99 latency: 2200 ms
Test stats: submitted: 2998 txn/s, committed: 2998 txn/s, expired: 0 txn/s, failed submission: 0 tnx/s, latency: 2284 ms, p99 latency: 5350 ms
Cooldown stats: submitted: 2995 txn/s, committed: 2979 txn/s, expired: 16 txn/s, failed submission: 0 tnx/s, latency: 1208 ms, p99 latency: 1650 ms

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/aptos-labs/aptos-core/3938)
<!-- Reviewable:end -->