@sitalkedia 2022-09-09
Hmm, interestingly this PR increased Forge performance TPS because of removing the TPS measurement during the startup and teardown period. I guess it is okay for us to measure the peak sustained TPS.